requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
decorator==4.4.2
ffmpeg-python==0.2.0
filelock==3.9.0
Flask==2.2.3
frozenlist==1.3.3
future==0.18.3
huggingface-hub==0.13.0
idna==3.4
imageio==2.26.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
llvmlite==0.39.1
lxml==4.9.2
MarkupSafe==2.1.2
more-itertools==9.1.0
moviepy==1.0.3
multidict==6.0.4
numba==0.56.4
numpy==1.23.5
openai==0.27.0
openai-whisper==20230307
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
proglog==0.1.10
pydub==0.25.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
six==1.16.0
tokenizers==0.13.2
torch==1.13.1
tqdm==4.65.0
transformers==4.26.1
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
